refactor: move DraggableResizableAsset to photobooth_ui
