+++ lib/assets/assets.dart
@@ -1,9 +1,8 @@
 import 'dart:typed_data';
-import 'dart:ui' as ui;
 
-import 'package:equatable/equatable.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/rendering.dart';
+import 'package:photobooth_ui/photobooth_ui.dart';
 
 abstract class Assets {
   static late final Asset dash;
@@ -27,15 +26,4 @@ Future<Asset> _loadAsset(String path) async {
   final imageBytes = await image.toByteData();
 
   return Asset(image: image, buffer: imageBytes!.buffer, bytes: bytes);
-}
-
-class Asset extends Equatable {
-  const Asset({required this.image, required this.buffer, required this.bytes});
-
-  final ui.Image image;
-  final ByteBuffer buffer;
-  final Uint8List bytes;
-
-  @override
-  List<Object?> get props => [image, buffer, bytes];
 }
