+++ lib/decoration/bloc/decoration_bloc.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:io_photobooth/assets/assets.dart';
+import 'package:photobooth_ui/photobooth_ui.dart';
 
 part 'decoration_event.dart';
 part 'decoration_state.dart';
