+++ lib/decoration/decoration.dart
@@ -1,3 +1,2 @@
 export 'bloc/decoration_bloc.dart';
 export 'view/decoration_page.dart';
-export 'widgets/widgets.dart';
