--- lib/decoration/widgets/widgets.dart
@@ -1,2 +1,0 @@
-export 'resizable_sticker.dart';
-export 'stickers_frame.dart';
