+++ packages/photobooth_ui/lib/photobooth_ui.dart
@@ -2,5 +2,6 @@ library photobooth_ui;
 
 export 'src/colors.dart';
 export 'src/layout/layout.dart';
+export 'src/models/models.dart';
 export 'src/theme.dart';
 export 'src/widgets/widgets.dart';
