+++ packages/photobooth_ui/lib/src/models/models.dart
@@ -1,0 +1,1 @@
+export 'asset.dart';
