+++ packages/photobooth_ui/lib/src/widgets/widgets.dart
@@ -1,2 +1,3 @@
+export 'draggable_resizable_asset.dart';
 export 'preview_image.dart';
 export 'responsive_layout_builder.dart';
